Updated models and model relationships

+++ Models/GazePoint.cs
@@ -2,11 +2,11 @@ namespace AccuGazer.API.Models
 {
     public class GazePoint
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public int X { get; set; }
         public int Y { get; set; }
 
-        public int MeasurementId { get; set; }
+        public long MeasurementId { get; set; }
         public Measurement Measurement { get; set; }
     }
 }
