+++ Models/Rectangle.cs
@@ -8,7 +8,7 @@ namespace AccuGazer.API.Models
         public int Width { get; set; }
         public int Height { get; set; }
 
-        public int MeasurementId { get; set; }
+        public long MeasurementId { get; set; }
         public Measurement Measurement { get; set; }
     }
 }
